:truck: Moved sections modules into Sections dir
