+++ src/Common/Components/Modules/Sections/Hero.scss
